Using cmake to compile the library
- Changing header files leaving separated the exported API from the
  internal functions.
