+++ include/api/error.h
@@ -6,7 +6,7 @@
  * Created at: Fri Feb 10 09:06:20 2017
  * Project: libv4l2
  *
- * Copyright (C) 2017 Rodrigo Freitas
+ * Copyright (C) 2017 Rodrigo Freitas All rights reserved
  *
  * This library is free software; you can redistribute it and/or
  * modify it under the terms of the GNU Lesser General Public
@@ -24,12 +24,12 @@
  * USA
  */
 
-#ifndef _LIBV4L2_V_ERROR_H
-#define _LIBV4L2_V_ERROR_H          1
+#ifndef _LIBV4L2_API_ERROR_H
+#define _LIBV4L2_API_ERROR_H
 
 #ifndef LIBV4L2_COMPILE
 # ifndef _LIBV4L2_H
-#  error "Never use <v_error.h> directly; include <libv4l2.h> instead."
+#  error "Never use <error.h> directly; include <libv4l2.h> instead."
 # endif
 #endif
 
@@ -69,11 +69,6 @@ enum v4l2_error_code {
 
     V4L2_ERROR_MAX_ERROR_CODE
 };
-
-#ifdef LIBV4L2_COMPILE
-void errno_clear(void);
-void errno_set(enum v4l2_error_code code);
-#endif
 
 /**
  * @name v4l2_get_last_error
