+++ include/api/image.h
@@ -6,7 +6,7 @@
  * Created at: Wed Feb 15 21:38:22 2017
  * Project: libv4l2
  *
- * Copyright (c) 2017 All rights reserved
+ * Copyright (C) 2017 Rodrigo Freitas All rights reserved
  *
  * This library is free software; you can redistribute it and/or
  * modify it under the terms of the GNU Lesser General Public
@@ -24,18 +24,13 @@
  * USA
  */
 
-#ifndef _LIBV4L2_V_IMAGE_H
-#define _LIBV4L2_V_IMAGE_H          1
+#ifndef _LIBV4L2_API_IMAGE_H
+#define _LIBV4L2_API_IMAGE_H
 
 #ifndef LIBV4L2_COMPILE
 # ifndef _LIBV4L2_H
-#  error "Never use <v_image.h> directly; include <libv4l2.h> instead."
+#  error "Never use <image.h> directly; include <libv4l2.h> instead."
 # endif
-#endif
-
-#ifdef LIBV4L2_COMPILE
-struct v4l2_image_s *new_v4l2_image_s(enum v4l2_image_format format,
-                                      int width, int height);
 #endif
 
 /**
