+++ include/api/info.h
@@ -6,7 +6,7 @@
  * Created at: Thu Feb 16 16:15:34 2017
  * Project: libv4l2
  *
- * Copyright (C) 2017 Rodrigo Freitas
+ * Copyright (C) 2017 Rodrigo Freitas All rights reserved
  *
  * This library is free software; you can redistribute it and/or
  * modify it under the terms of the GNU Lesser General Public
@@ -24,12 +24,12 @@
  * USA
  */
 
-#ifndef _LIBV4L2_V_INFO_H
-#define _LIBV4L2_V_INFO_H          1
+#ifndef _LIBV4L2_API_INFO_H
+#define _LIBV4L2_API_INFO_H
 
 #ifndef LIBV4L2_COMPILE
 # ifndef _LIBV4L2_H
-#  error "Never use <v_info.h> directly; include <libv4l2.h> instead."
+#  error "Never use <info.h> directly; include <libv4l2.h> instead."
 # endif
 #endif
