+++ include/internal/mmap.h
@@ -6,7 +6,7 @@
  * Created at: Wed Feb 15 21:38:26 2017
  * Project: libv4l2
  *
- * Copyright (c) 2017 All rights reserved
+ * Copyright (C) 2017 Rodrigo Freitas All rights reserved
  *
  * This library is free software; you can redistribute it and/or
  * modify it under the terms of the GNU Lesser General Public
@@ -24,14 +24,8 @@
  * USA
  */
 
-#ifndef _LIBV4L2_V_MMAP_H
-#define _LIBV4L2_V_MMAP_H          1
-
-#ifndef LIBV4L2_COMPILE
-# ifndef _LIBV4L2_H
-#  error "Never use <v_mmap.h> directly; include <libv4l2.h> instead."
-# endif
-#endif
+#ifndef _LIBV4L2_INTERNAL_MMAP_H
+#define _LIBV4L2_INTERNAL_MMAP_H
 
 int mmap_init(struct v4l2_s *v4l2);
 int mmap_uninit(struct v4l2_s *v4l2);
