+++ misc/libv4l2.sym
